refactor(core): one home for selection dedupe semantics
combine() re-implemented makeSelection's prev/equals/notify loop — the
third copy after the machine's own and the watchers'. Extract
makeSelection(selector, attach) to selection.ts; the machine attaches
to its bus, compose attaches via the group registry.
